feat(decl): add test ID support
Add test ID support to enable unique identification of the process
chain's leaf process. The test ID is stored in the environment of
each process chain's process, except the root process. The root
process generate a test ID in the form <ignorePrefix><testUID>.
Each process in the chain but the one before the leaf passes the
untouched test ID to the child. The one before the leaf passes
the test ID without the <ignorePrefix> to the leaf.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>
Co-authored-by: Aldo Lacuku <aldo@lacuku.eu>