feat(decl/loader): set up two-steps unmarshaling via mapstructure
Unmarshal tests description in two steps: the first step performs
the YAML-specific conversion of the description into a generic
representation; the second step takes the generic representation
and uses mapstructure to place the content in the right structures.
Splitting the unmarshalling in two steps enables performing the
validation logic (through JSON schemas) before placing the content
in the corresponding structures.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>
Co-authored-by: Aldo Lacuku <aldo@lacuku.eu>