feat(decl/loader): add support for generating documentation node tree
Add support for generating a documentation node tree from JSON
schema. The main introduced schema.Describe function generates a
node tree by taking into account the provided node path and an
optional list of enum requirements. An enum requirement specifies
the node path to an enumerated value in the node tree and the
corresponding value to set: setting a value for an enumerated node
can "unlock" further nodes in the tree.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>
Co-authored-by: Aldo Lacuku <aldo@lacuku.eu>