feat(decl/loader): add test cases support
Add test cases support by providing the capability to specify
template parameters in tests description. A template parameter is
specified using the syntax %{item.<keyName>} in place of a concrete
value. A test description allows to specify test case
specifications using the `cases` keyword. A test case specification
can use one the two supported strategy: `vector` and `matrix`.
The values provided in a test case spec are interpreted differently
based on the strategy employed:
- if the strategy is set to `vector`, values are simply used to fill
  template parameters matching the corresponding value keys
- if the strategy is set to `matrix`, values for each key must be
  lists; combinations of all lists values are used to generate a
  number of test cases equal to the product of the cardinalities of
  all lists
A test template is not instantiated if it doesn't specify any case.
All case specifications in a test must specify the same set of value
keys.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>