feat(suite/step/syscalls): add arguments constraints support
Introduce mutually-exclusive-arguments, one-required-argument and
required-together-arguments constraints.
Mutually-exclusive-arguments constraint allows to mandate the user to
specify a value for at most one of a set of arguments.
One-required-argument constraint allows to mandate the user to specify
a value for at least one of a set of arguments.
Required-together-arguments constraints allows to mandate the user to
specify a value for each argument of a set of arguments if it provides
a value for at least one of them.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>