fix(suite/step/syscalls): align `read` and `write` step args after run
Upon successful execution, update buffers and lengths arguments to
their new value. This is especially important when buffers are created
on the fly with the specified length, or when lengths default to the
buffers length. This is required in order to enable subsequent steps
to correctly see the updated values.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>